requirements.txt
Flask==3.1.0
psycopg2-binary
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.4.1
Flask-CORS==5.0.0
Flask-Limiter==3.8.0
psycopg2-binary==2.9.11
SQLAlchemy==2.0.38
twilio==9.4.6
firebase-admin==6.6.0
gunicorn==23.0.0
python-dotenv==1.0.1
marshmallow==3.23.2
bcrypt==4.2.1
PyJWT==2.10.1
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pytest==8.0.0
pandas>=2.0.0
pytz==2024.1
